refactor: 공통 사용 convert 로직 분리
