+++ src/main/java/com/bitharmony/comma/album/album/controller/AlbumController.java
@@ -1,5 +1,6 @@
 package com.bitharmony.comma.album.album.controller;
 
+import com.bitharmony.comma.album.album.util.AlbumConvertUtil;
 import java.security.Principal;
 
 import org.springframework.data.domain.Page;
@@ -38,6 +39,7 @@ public class AlbumController {
 
 	private final AlbumService albumService;
 	private final AlbumLikeService albumLikeService;
+	private final AlbumConvertUtil albumConvertUtil;
 	private final MemberService memberService;
 
 	@PostMapping("/release")
@@ -52,14 +54,14 @@ public class AlbumController {
 		}
 
 		Album album = albumService.release(request, member, musicImageFile);
-		return GlobalResponse.of("200", albumToResponseDto(album));
+		return GlobalResponse.of("200", albumConvertUtil.albumToResponseDto(album));
 	}
 
 	@GetMapping("/detail/{id}")
 	public GlobalResponse getAlbum(@PathVariable long id) {
 		Album album = albumService.getAlbumById(id);
 
-		return GlobalResponse.of("200", albumToResponseDto(album));
+		return GlobalResponse.of("200", albumConvertUtil.albumToResponseDto(album));
 	}
 
 	@GetMapping("/{username}")
@@ -92,7 +94,7 @@ public class AlbumController {
 		}
 
 		Album editedAlbum = albumService.edit(request, album, musicImageFile);
-		return GlobalResponse.of("200", albumToResponseDto(editedAlbum));
+		return GlobalResponse.of("200", albumConvertUtil.albumToResponseDto(editedAlbum));
 	}
 
 	@DeleteMapping("/{id}")
@@ -144,25 +146,5 @@ public class AlbumController {
 		albumLikeService.like(member, album);
 		return GlobalResponse.of("200");
 	}
-
-	private AlbumResponse albumToResponseDto(Album album) {
-		album = album.toBuilder()
-			//.filePath(albumService.getAlbumFileUrl(album.getFilePath()))
-			.imagePath(albumService.getAlbumImageUrl(album.getImagePath()))
-			.build();
-
-		return AlbumResponse.builder()
-			.id(album.getId())
-			.albumname(album.getAlbumname())
-			.genre(album.getGenre())
-			.license(album.isLicense())
-			.licenseDescription(album.getLicenseDescription())
-			.imgPath(album.getImagePath())
-			.filePath(album.getFilePath())
-			.permit(album.isPermit())
-			.price(album.getPrice())
-			.artistNickname(album.getMember().getNickname())
-			.artistUsername(album.getMember().getUsername())
-			.build();
-	}
+	
 }
